requirements.txt
requests
pandas
numpy
boto3
python-dotenv
